# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Import/Export management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard:
    get:
      operationId: getDashboard
      tags: [rapports]
      summary: Get dashboard statistics
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardData"
  /bilan-journalier:
    get:
      operationId: getBilanJournalier
      tags: [rapports]
      summary: Get daily report
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BilanJournalier"
components:
  schemas:
...
    Colis:
      type: object
      properties:
        id:
          type: integer
        arrivageId:
          type: integer
        codeColisReel:
          type: string
        codeColisComplet:
          type: string
        type:
          type: string
          enum: [individuel, mixte]
        poids:
          type: number
        montant:
          type: number
        dateCreation:
          type: string
          format: date-time
      required:
        - id
        - arrivageId
        - codeColisReel
        - codeColisComplet
        - type
        - poids
        - montant
        - dateCreation
    ColisProprietaire:
      type: object
      properties:
        id:
          type: integer
        colisId:
          type: integer
        clientId:
          type: integer
        clientNom:
          type: string
        clientTelephone:
          type: string
        poids:
          type: number
        montantDu:
          type: number
        montantPaye:
          type: number
        solde:
          type: number
      required:
        - id
        - colisId
        - clientId
        - clientNom
        - clientTelephone
        - poids
        - montantDu
        - montantPaye
        - solde
    ColisDetail:
      allOf:
        - $ref: "#/components/schemas/Colis"
        - type: object
          properties:
            proprietaires:
              type: array
              items:
                $ref: "#/components/schemas/ColisProprietaire"
            arrivageDate:
              type: string
              format: date
          required:
            - proprietaires
            - arrivageDate
...
    DetteDetail:
      type: object
      properties:
        colisProprietaireId:
          type: integer
        clientId:
          type: integer
        clientNom:
          type: string
        clientTelephone:
          type: string
        colisCode:
          type: string
        arrivageDate:
          type: string
          format: date
        montantDu:
          type: number
        montantPaye:
          type: number
        solde:
          type: number
      required:
        - colisProprietaireId
        - clientId
        - clientNom
        - clientTelephone
        - colisCode
        - arrivageDate
        - montantDu
        - montantPaye
        - solde
    DashboardData:
      type: object
      properties:
        today:
          type: object
          properties:
            nbColis:
              type: integer
            poidsTotal:
              type: number
            montantTotal:
              type: number
            montantEncaisse:
              type: number
            dettesTotal:
              type: number
          required:
            - nbColis
            - poidsTotal
            - montantTotal
            - montantEncaisse
            - dettesTotal
        evolutionColis:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              nbColis:
                type: integer
              montant:
                type: number
            required:
              - date
              - nbColis
              - montant
        topClients:
          type: array
          items:
            type: object
            properties:
              clientId:
                type: integer
              clientNom:
                type: string
              totalDu:
                type: number
              solde:
                type: number
            required:
              - clientId
              - clientNom
              - totalDu
              - solde
        totalDettesGlobal:
          type: number
      required:
        - today
        - evolutionColis
        - topClients
        - totalDettesGlobal
    BilanJournalier:
      type: object
      properties:
        date:
          type: string
          format: date
        nbColisIndividuels:
          type: integer
        nbColisMixtes:
          type: integer
        poidsIndividuels:
          type: number
        poidsMixtes:
          type: number
        montantTotal:
          type: number
        montantEncaisse:
          type: number
        montantDu:
          type: number
        dettes:
          type: array
          items:
            $ref: "#/components/schemas/DetteDetail"
        colis:
          type: array
          items:
            $ref: "#/components/schemas/ColisDetail"
      required:
        - date
        - nbColisIndividuels
        - nbColisMixtes
        - poidsIndividuels
        - poidsMixtes
        - montantTotal
        - montantEncaisse
        - montantDu
        - dettes
        - colis