# lib/api-zod/src/generated/types/bilanJournalier.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Import/Export management API
 * OpenAPI spec version: 0.1.0
 */
import type { ColisDetail } from "./colisDetail";
import type { DetteDetail } from "./detteDetail";

export interface BilanJournalier {
  date: Date;
  nbColisIndividuels: number;
  nbColisMixtes: number;
  poidsIndividuels: number;
  poidsMixtes: number;
  montantTotal: number;
  montantEncaisse: number;
  montantDu: number;
  dettes: DetteDetail[];
  colis: ColisDetail[];
}